Timezone & Locale stored in group preferences.
